Use usize instead of u32

+++ src/editor/mod.rs
@@ -6,7 +6,7 @@ pub struct Editor {
     core: Core,
 }
 
-pub fn new(buffer: String, line: u32, column: u32) -> Result<Editor, String> {
+pub fn new(buffer: String, line: usize, column: usize) -> Result<Editor, String> {
     match core::new(buffer, line, column) {
         Ok(editor) => Ok(Editor { core: editor }),
         Err(err) => Err(err),
@@ -14,23 +14,23 @@ pub fn new(buffer: String, line: u32, column: u32) -> Result<Editor, String> {
 }
 
 impl Editor {
-    pub fn move_right(&mut self, n: u32) {
+    pub fn move_right(&mut self, n: usize) {
         self.core.move_right(n);
     }
 
-    pub fn move_left(&mut self, n: u32) {
+    pub fn move_left(&mut self, n: usize) {
         self.core.move_left(n);
     }
 
-    pub fn move_up(&mut self, n: u32) {
+    pub fn move_up(&mut self, n: usize) {
         self.core.move_up(n);
     }
 
-    pub fn move_down(&mut self, n: u32) {
+    pub fn move_down(&mut self, n: usize) {
         self.core.move_down(n);
     }
 
-    pub fn insert_at(&mut self, ch: char, line: u32, column: u32) {
+    pub fn insert_at(&mut self, ch: char, line: usize, column: usize) {
         self.core.insert_at(ch, line, column);
     }
 
@@ -38,11 +38,11 @@ impl Editor {
         self.core.buffer()
     }
 
-    pub fn line(&self) -> u32 {
+    pub fn line(&self) -> usize {
         self.core.line()
     }
 
-    pub fn column(&self) -> u32 {
+    pub fn column(&self) -> usize {
         self.core.column()
     }
 }
